device.py_monitor_message_transmission.py/requirements.txt
Flask
azure-iot-device
json
traceback
time